+++ ledger/src/error.rs
@@ -13,12 +13,11 @@
 
 use crate::dust::{DustGenerationInfo, DustNullifier, DustRegistration, DustSpend, InitialNonce};
 use crate::error::coin::UserAddress;
-use crate::structure::MAX_SUPPLY;
 use crate::structure::{ClaimKind, ContractOperationVersion, Utxo, UtxoOutput, UtxoSpend};
+use crate::structure::{MAX_SUPPLY, SignatureVerifyingKey};
 use base_crypto::cost_model::CostDuration;
 use base_crypto::fab::{Alignment, Value};
 use base_crypto::hash::HashOutput;
-use base_crypto::schnorr::VerifyingKey;
 use base_crypto::time::Timestamp;
 use coin_structure::coin::{self, Commitment, Nullifier, PublicAddress, TokenType};
 use coin_structure::contract::ContractAddress;
@@ -485,7 +484,7 @@ pub enum MalformedTransaction<D: DB> {
         validity_end: Timestamp,
     },
     MultipleDustRegistrationsForKey {
-        key: VerifyingKey,
+        key: SignatureVerifyingKey,
     },
     InsufficientDustForRegistrationFee {
         registration: Box<DustRegistration<(), D>>,