+++ CHANGELOG.md
@@ -8,6 +8,7 @@ with `zswap` being tracked in [Changelog Zswap](./CHANGELOG_zswap.md).
 
 - feat: add explicit price floor, denominated in full blocks, and governed by
   ledger parameters.
+- feat: add support for ECDSA signatures
 - feat: add `UnlockToTreasury` system transaction, moving funds from the locked
   pool to the treasury.
 - fix: correctly exclude the identity point during coin ciphertext decryption